Fix URL link mechanism for Red Hat Errata
Previously, public description of each erratum was available under

    https://rhn.redhat.com/errata/

at some point in time this has changed to

    https://access.redhat.com/errata/

and since not all links redirect properly from old to new, we have to fix
OpenSCAP end as well.

This enables people reviewing XCCDF policy guide or XCCDF result report to click
on RH[SBE]A-YYYY-ABCD acronym and be redirected to public description of the
given erratum.