tests/probes/process58/all.sh
Check if process58 probe is available before trying to test it. This
fixes an issue where if a system did not support the process58 probe,
these tests would fail instead of being skipped. An example of this is
FreeBSD which does not have libcap, and therefore does not support the
process58 probe.